refactor React Router integration
We no longer use `match` to do preemptive routing. This was a hack anyways. `Router` now hooks into
the `history` listener and has full access to the actual `routerState` before `render` is called.
With this change, `RouterContext` is decoupled from the `Router` and is just a stateless container
that is responsible for the rendering of the `navigationTree`.  This will make testing much easier.